refactor: move icons from asset to their related component folders
