+++ app/components/Block/ImageBlock.jsx
@@ -4,11 +4,11 @@ import BlockDescription from "../BlockDescription/BlockDescription";
 
 import './block.scss'
 // Icons
-import BlockIcon from "../../asset/collection/images.svg";
-import TrashIcon from "../../asset/collection/trash-alt.svg";
-import AngleIcon from "../../asset/collection/angle-up.svg";
-import MarkIcon from "../../asset/collection/bookmark.svg";
-import MarkFullIcon from "../../asset/collection/bookmark-full.svg";
+import BlockIcon from "./icons/images.svg";
+import TrashIcon from "./icons/trash-alt.svg";
+import AngleIcon from "./icons/angle-up.svg";
+import MarkIcon from "./icons/bookmark.svg";
+import MarkFullIcon from "./icons/bookmark-full.svg";
 
 export default function ImageBlock(props) {
     const scaleid = "scale_" + props.block.timestamp;
